feat: api, sanctum, completed ecommerce api

+++ app/Http/Controllers/ProductController.php
@@ -2,9 +2,9 @@
 
 namespace App\Http\Controllers;
 
+use App\Http\Requests\StoreProductRequest;
+use App\Http\Requests\UpdateProductRequest;
 use App\Models\Product;
-use App\Models\User;
-use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 
 class ProductController extends Controller
@@ -32,19 +32,10 @@ class ProductController extends Controller
         return redirect('/login');
     }
 
-    public function store(Request $request)
+    public function store(StoreProductRequest $request)
     {
         if (Auth::check()) {
-            $validated =   $request->validate(
-                [
-                    'name' => ['required', 'max:255'],
-                    'description' => ['required', 'max:256'],
-                    'price' => ['required', 'numeric', 'gt:0'],
-                    'quantity' => ['required', 'numeric', 'gt:0'],
-                    'type' => ['required', 'in:physical,digital'],
-                ]
-
-            );
+            $validated =   $request->validated();
 
             if (Product::create([...$validated, 'user_id' => Auth::id()])) {
                 return redirect('/products')->with('success', 'Product created successfully!');;
@@ -81,16 +72,11 @@ class ProductController extends Controller
         return redirect('/login');
     }
 
-    public function update(string $id, Request $request)
+    public function update(string $id, UpdateProductRequest $request)
     {
-        $validated = $request->validate([
-            'name' => ['string', 'max:255'],
-            'description' => ['string', 'max:255'],
-            'price' => ['numeric', 'gt:0'],
-            'quantity' => ['numeric', 'gt:0'],
-            'type' => ['in:physical,digital']
-        ]);
 
+
+        $validated = $request->validated();
 
         if (Product::where('product_id', $id)->update(
             $validated
@@ -108,6 +94,10 @@ class ProductController extends Controller
         }
 
         $product = Product::findOrFail($id);
+
+        if ($product->product_id != Auth::id()) {
+            abort(403, "un authorized access");
+        }
 
         if ($product->delete()) {
             return redirect('/products/user-product');
